rename product name
