+++ settings.gradle.kts
@@ -3,4 +3,4 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
-rootProject.name = "mercari"
+rootProject.name = "libermo"
