+++ src/main/kotlin/jp/ac/titech/itsp/libermo/configuration/SwaggerConfiguration.kt
@@ -1,18 +1,15 @@
-package jp.ac.titech.itsp.mercari.configuration
+package jp.ac.titech.itsp.libermo.configuration
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import springfox.documentation.swagger2.annotations.EnableSwagger2
 import springfox.documentation.builders.ApiInfoBuilder
-import springfox.documentation.service.ApiInfo
 import springfox.documentation.builders.PathSelectors
+import springfox.documentation.service.ApiInfo
 import springfox.documentation.spi.DocumentationType
 import springfox.documentation.spring.web.plugins.Docket
-import springfox.documentation.swagger.web.UiConfigurationBuilder
 import springfox.documentation.swagger.web.UiConfiguration
-
-
-
+import springfox.documentation.swagger.web.UiConfigurationBuilder
+import springfox.documentation.swagger2.annotations.EnableSwagger2
 
 @Configuration
 @EnableSwagger2
@@ -26,8 +23,8 @@ class SwaggerConfiguration {
         .apiInfo(apiInfo())
 
     private fun apiInfo(): ApiInfo = ApiInfoBuilder()
-        .title("titech-mercari")
-        .description("titech-mercari api")
+        .title("libermo")
+        .description("libermo api document")
         .version("0.0.1-SNAPSHOT")
         .build()
