+++ src/main/kotlin/jp/ac/titech/itsp/libermo/configuration/WebSecurityConfiguration.kt
@@ -1,4 +1,4 @@
-package jp.ac.titech.itsp.mercari.configuration
+package jp.ac.titech.itsp.libermo.configuration
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
