+++ src/main/kotlin/jp/ac/titech/itsp/libermo/exceptions/ForbiddenException.kt
@@ -1,3 +1,3 @@
-package jp.ac.titech.itsp.mercari.exceptions
+package jp.ac.titech.itsp.libermo.exceptions
 
 class ForbiddenException(message: String) : RuntimeException(message)
