+++ src/main/kotlin/jp/ac/titech/itsp/libermo/models/Item.kt
@@ -1,4 +1,4 @@
-package jp.ac.titech.itsp.mercari.models
+package jp.ac.titech.itsp.libermo.models
 
 import javax.persistence.*
