+++ src/main/kotlin/jp/ac/titech/itsp/libermo/models/User.kt
@@ -1,4 +1,4 @@
-package jp.ac.titech.itsp.mercari.models
+package jp.ac.titech.itsp.libermo.models
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import javax.persistence.*
