+++ src/main/kotlin/jp/ac/titech/itsp/libermo/repositories/ItemRepository.kt
@@ -1,9 +1,8 @@
-package jp.ac.titech.itsp.mercari.repositories
+package jp.ac.titech.itsp.libermo.repositories
 
-import jp.ac.titech.itsp.mercari.models.Item
+import jp.ac.titech.itsp.libermo.models.Item
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import java.util.*
 
 @Repository
 interface ItemRepository : JpaRepository<Item, Long> {
