+++ src/main/kotlin/jp/ac/titech/itsp/libermo/repositories/UserRepository.kt
@@ -1,6 +1,6 @@
-package jp.ac.titech.itsp.mercari.repositories
+package jp.ac.titech.itsp.libermo.repositories
 
-import jp.ac.titech.itsp.mercari.models.User
+import jp.ac.titech.itsp.libermo.models.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
