+++ src/main/kotlin/jp/ac/titech/itsp/libermo/services/ItemService.kt
@@ -1,12 +1,11 @@
-package jp.ac.titech.itsp.mercari.services
+package jp.ac.titech.itsp.libermo.services
 
 import javassist.NotFoundException
-import jp.ac.titech.itsp.mercari.exceptions.ForbiddenException
-import jp.ac.titech.itsp.mercari.exceptions.IllegalStateException
-import jp.ac.titech.itsp.mercari.models.Item
-import jp.ac.titech.itsp.mercari.models.ItemState
-import jp.ac.titech.itsp.mercari.models.User
-import jp.ac.titech.itsp.mercari.repositories.ItemRepository
+import jp.ac.titech.itsp.libermo.exceptions.ForbiddenException
+import jp.ac.titech.itsp.libermo.exceptions.IllegalStateException
+import jp.ac.titech.itsp.libermo.models.Item
+import jp.ac.titech.itsp.libermo.models.ItemState
+import jp.ac.titech.itsp.libermo.repositories.ItemRepository
 import org.springframework.stereotype.Service
 
 @Service
